[ai][go] velocity lidar namespace: fold in pcap-analyse
Extract the pcap-analyse engine into internal/lidar/pcapanalyse (Run(Config) int) and add an internal/cmd/lidar namespace applet that parses flags into a FlagSet and calls the engine, with a !pcap stub so the default toolchain stays green. Register 'velocity lidar pcap-analyse' in the root router and reduce cmd/tools/pcap-analyse to a thin wrapper over the shared applet. Output and flags are byte-identical to the standalone tool; root routing is covered in root_test.go. Implements docs/plans/velocity-lidar-namespace-plan.md.
