+++ internal/cmd/root/root.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	"github.com/banshee-data/velocity.report/internal/cmd/device"
+	"github.com/banshee-data/velocity.report/internal/cmd/lidar"
 	"github.com/banshee-data/velocity.report/internal/cmd/server"
 	"github.com/banshee-data/velocity.report/internal/cmd/tune"
 	"github.com/banshee-data/velocity.report/internal/version"
@@ -15,6 +16,7 @@ import (
 var (
 	serverMain   = server.Main
 	deviceMain   = device.Main
+	lidarMain    = lidar.Main
 	tuneMain     = tune.Main
 	printVersion = version.Print
 )
@@ -27,6 +29,7 @@ Usage:
 Namespaces:
   serve     Run the radar/LiDAR server
   device    On-device lifecycle: check, upgrade, rollback, backup, status, tailscale
+  lidar     LiDAR capture diagnostics: pcap-analyse
   data      Database operations: migrate, transits, sql
   report    Generate PDF reports: pdf
   tune      Parameter tuning: sweep
@@ -60,6 +63,8 @@ func Dispatch(prog string, args []string) int {
 		return serverMain(args[1:])
 	case "device":
 		return deviceMain(args[1:])
+	case "lidar":
+		return lidarMain(args[1:])
 	case "data":
 		if len(args) >= 2 && (args[1] == "migrate" || args[1] == "transits" || args[1] == "sql") {
 			return serverMain(args[1:])
