+++ internal/lidar/pcapanalyse/stats_test.go
@@ -1,7 +1,7 @@
 //go:build pcap
 // +build pcap
 
-package main
+package pcapanalyse
 
 import (
 	"math"
